requirements.txt
colorgram.py==1.2.0
imageio==2.27.0
lazy_loader==0.2
mysql-connector-python==8.0.33
networkx==3.1
numpy==1.24.2
opencv-contrib-python==4.7.0.72
packaging==23.0
Pillow==9.5.0
protobuf==3.20.3
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.10.1
tifffile==2023.3.21
